add confirm when delete

+++ admin/includes/view_all_posts.php
@@ -91,7 +91,7 @@ if (isset($_POST['checkBoxArray'])) {
         echo "<td>{$post_tags}</td>";
         echo "<td>{$post_comment_count}</td>";
         echo "<td>{$post_date}</td>";
-        echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+        echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
         echo "<td><a href='posts.php?source=edit_post&edit_id={$post_id}'>Edit</></td>";
         echo "</tr>";
     }
